Made AddClient.vue more responsive and refactored the name of AddJobList.vue to AddJob.vue and changed the routing, also made changes to modals view
